Implimented CentralLogger using Timber and Added JavaDocs to all Files.
