--- app/src/main/java/com/starstruckstech/pgpodandroid/LogAdapter.kt
@@ -1,26 +1,0 @@
-package com.starstruckstech.pgpodandroid
-
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import android.widget.TextView
-import androidx.recyclerview.widget.RecyclerView
-
-class LogAdapter(private val logs: List<LogEntry>) : RecyclerView.Adapter<LogAdapter.LogViewHolder>() {
-
-    class LogViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView) {
-        val textViewLog: TextView = itemView.findViewById(R.id.textViewLog)
-    }
-
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): LogViewHolder {
-        val view = LayoutInflater.from(parent.context).inflate(R.layout.item_log, parent, false)
-        return LogViewHolder(view)
-    }
-
-    override fun onBindViewHolder(holder: LogViewHolder, position: Int) {
-        val logEntry = logs[position]
-        holder.textViewLog.text = "${logEntry.timestamp}: [${logEntry.level}] ${logEntry.tag} - ${logEntry.message}"
-    }
-
-    override fun getItemCount(): Int = logs.size
-}
