--- app/src/main/java/com/starstruckstech/pgpodandroid/LogEntry.kt
@@ -1,8 +1,0 @@
-package com.starstruckstech.pgpodandroid
-
-data class LogEntry(
-    val level: LogLevel,
-    val tag: String,
-    val message: String,
-    val timestamp: Long = System.currentTimeMillis()
-)
