--- app/src/main/java/com/starstruckstech/pgpodandroid/LogLevel.kt
@@ -1,5 +1,0 @@
-package com.starstruckstech.pgpodandroid
-
-enum class LogLevel {
-    DEBUG, INFO, WARN, ERROR
-}
