+++ app/src/main/java/com/starstruckstech/pgpodandroid/MainActivity.kt
@@ -1,6 +1,7 @@
 package com.starstruckstech.pgpodandroid
 
 import android.os.Bundle
+import android.os.PersistableBundle
 import androidx.activity.enableEdgeToEdge
 import androidx.appcompat.app.AppCompatActivity
 import com.google.firebase.FirebaseApp
@@ -10,7 +11,10 @@ import androidx.compose.ui.platform.ComposeView
 import androidx.core.view.ViewCompat
 import androidx.core.view.WindowInsetsCompat
 import androidx.lifecycle.viewmodel.compose.viewModel
+import com.starstruckstech.pgpodandroid.central_logger.CentralLogger
+import com.starstruckstech.pgpodandroid.central_logger.LogLevel
 import okhttp3.internal.platform.android.BouncyCastleSocketAdapter.Companion.factory
+import timber.log.Timber
 
 class MainActivity : AppCompatActivity() {
 
@@ -21,9 +25,8 @@ class MainActivity : AppCompatActivity() {
 
         FirebaseApp.initializeApp(this)
 
-        CentralLogger.log(LogLevel.INFO, "MainActivity", "App started")
-    }
-}
+        Timber.i("App started")
+
         ViewCompat.setOnApplyWindowInsetsListener(findViewById(R.id.main)) { v, insets ->
             val systemBars = insets.getInsets(WindowInsetsCompat.Type.systemBars())
             v.setPadding(systemBars.left, systemBars.top, systemBars.right, systemBars.bottom)
@@ -42,6 +45,8 @@ class MainActivity : AppCompatActivity() {
 
         }
     }
-
+    override fun onStart() {
+        super.onStart()
+        Timber.i("onStart")
+    }
 }
-    }}}
