+++ app/src/main/java/com/starstruckstech/pgpodandroid/central_logger/LogLevel.kt
@@ -1,0 +1,9 @@
+package com.starstruckstech.pgpodandroid.central_logger
+
+/**
+ * This enum class is used to define the log levels
+ * DEBUG, INFO, WARN, ERROR
+ */
+enum class LogLevel {
+    DEBUG, INFO, WARN, ERROR
+}
