--- app/src/main/res/layout/activity_log_viewer.xml
@@ -1,11 +1,0 @@
-<LinearLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical">
-
-    <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/recyclerViewLogs"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"/>
-</LinearLayout>
